+++ src/Test.FastDisconnect/Program.cs
@@ -5,17 +5,16 @@ namespace Test.FastDisconnect
 {
     class Program
     {
-        static WatsonTcpClient _Client = null;
-
         static void Main(string[] args)
         {
             try
             {
-                _Client = new WatsonTcpClient("127.0.0.1", 9000);
-                _Client.Events.MessageReceived += MessageReceived;
-                _Client.Connect();
-                _Client.Send("Hello!");
-                _Client.Dispose();
+                using (WatsonTcpClient client = new WatsonTcpClient("127.0.0.1", 9000))
+                {
+                    client.Events.MessageReceived += MessageReceived;
+                    client.Connect();
+                    client.Send("Hello!");
+                }
             }
             catch (Exception e)
             {