+++ src/WatsonTcp/DefaultSerializationHelper.cs
@@ -1,5 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Collections.Specialized;
+using System.Linq;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 
@@ -15,6 +17,9 @@ namespace WatsonTcp
         #endregion
 
         #region Private-Members
+
+        private ExceptionConverter<Exception> _ExceptionConverter = new ExceptionConverter<Exception>();
+        private NameValueCollectionConverter _NameValueCollectionConverter = new NameValueCollectionConverter();
 
         #endregion
 
@@ -53,13 +58,22 @@ namespace WatsonTcp
         {
             if (obj == null) return null;
 
+            JsonSerializerOptions options = new JsonSerializerOptions();
+            options.DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingNull;
+
+            // see https://github.com/dotnet/runtime/issues/43026
+            options.Converters.Add(_ExceptionConverter);
+            options.Converters.Add(_NameValueCollectionConverter);
+
             if (!pretty)
             {
-                return JsonSerializer.Serialize(obj, new JsonSerializerOptions { WriteIndented = false, DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingDefault });
+                options.WriteIndented = false;
+                return JsonSerializer.Serialize(obj, options);
             }
             else
             {
-                return JsonSerializer.Serialize(obj, new JsonSerializerOptions { WriteIndented = true, DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingDefault });
+                options.WriteIndented = true;
+                return JsonSerializer.Serialize(obj, options);
             }
         }
 
@@ -81,6 +95,66 @@ namespace WatsonTcp
         #endregion
 
         #region Private-Methods
+
+        #endregion
+
+        #region Private-Classes
+
+        private class ExceptionConverter<TExceptionType> : JsonConverter<TExceptionType>
+        {
+            public override bool CanConvert(Type typeToConvert)
+            {
+                return typeof(Exception).IsAssignableFrom(typeToConvert);
+            }
+
+            public override TExceptionType Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
+            {
+                throw new NotSupportedException("Deserializing exceptions is not allowed");
+            }
+
+            public override void Write(Utf8JsonWriter writer, TExceptionType value, JsonSerializerOptions options)
+            {
+                var serializableProperties = value.GetType()
+                    .GetProperties()
+                    .Select(uu => new { uu.Name, Value = uu.GetValue(value) })
+                    .Where(uu => uu.Name != nameof(Exception.TargetSite));
+
+                if (options.DefaultIgnoreCondition == JsonIgnoreCondition.WhenWritingNull)
+                {
+                    serializableProperties = serializableProperties.Where(uu => uu.Value != null);
+                }
+
+                var propList = serializableProperties.ToList();
+
+                if (propList.Count == 0)
+                {
+                    // Nothing to write
+                    return;
+                }
+
+                writer.WriteStartObject();
+
+                foreach (var prop in propList)
+                {
+                    writer.WritePropertyName(prop.Name);
+                    JsonSerializer.Serialize(writer, prop.Value, options);
+                }
+
+                writer.WriteEndObject();
+            }
+        }
+
+        private class NameValueCollectionConverter : JsonConverter<NameValueCollection>
+        {
+            public override NameValueCollection Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options) => throw new NotImplementedException();
+
+            public override void Write(Utf8JsonWriter writer, NameValueCollection value, JsonSerializerOptions options)
+            {
+                var val = value.Keys.Cast<string>()
+                    .ToDictionary(k => k, k => string.Join(", ", value.GetValues(k)));
+                System.Text.Json.JsonSerializer.Serialize(writer, val);
+            }
+        }
 
         #endregion
     }